Ensure DB is created when running
If `dbContext.Database.Migrate()` is run unconditionally, and the running user is pointing the application to an already-existing database, an attempt will be made to migrate an already-migrated database, which throws an error. Now, with `Ensure.Created()`, a check is made to see if a non-empty database already exists where the application is pointing; if so, the migrations aren't applied.

There is the possibility that the user is using a pre-existing database that isn't empty but that isn't quite up to date. In that case, the user will just have to drop their database and try again.