Fix migration error bug
With how frequently we reset our migrations, it's necessary to fully reset our local databases. However, `.GetPendingMigrations()` keeps returning true since the "Initial" migrations are timestamped. So, EF tries to apply an Initial migration to an already-initialized database, which causes problems. Adding `.EnsureDeleted` into the mix makes it so if our local DBs are in any way out of sync with the current migrations, they just get wiped and re-initialized.